aoc2023/pyproject.toml
[tool.poetry]
name = "aoc2023"
version = "0.1.0"
description = ""
authors = ["Baptiste Fontaine <b@ptistefontaine.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
advent-of-code-data = "^2.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
